Separate addButton in new framework mapping flow
