+++ app/views/AnalyticalFramework/AssistedTagging/WidgetTagList/Matrix1dTagInput/index.tsx
@@ -7,17 +7,8 @@ import {
     randomString,
 } from '@togglecorp/fujs';
 import {
-    List,
     ListView,
-    Tag,
-    QuickActionButton,
-    useBooleanState,
-    SelectInput,
 } from '@the-deep/deep-ui';
-import {
-    IoCloseOutline,
-    IoAdd,
-} from 'react-icons/io5';
 
 import {
     Matrix1dWidget,
@@ -31,137 +22,9 @@ import FrameworkTagRow from '../../FrameworkTagRow';
 
 import styles from './styles.css';
 
-const mappingKeySelector = (mapping: Matrix1dMappingsItem) => mapping.tag;
-const predictionKeySelector = (prediction: PredictionTag) => prediction.id;
-const predictionLabelSelector = (prediction: PredictionTag) => prediction.name;
-const predictionGroupKeySelector = (prediction: PredictionTag) => prediction.group ?? 'Misc';
-
-interface TagRowProps {
-    title: string;
-    cellKey: string;
-    onMappingRemoveClick: (cellKey: string, tagKey: string) => void;
-    onMappingAddClick: (cellKey: string, tagKey: string) => void;
-    mappings: Matrix1dMappingsItem[] | undefined;
-    predictionTagsById: Record<string, PredictionTag> | undefined;
-    disabled?: boolean;
-}
-
-function TagRow(props: TagRowProps) {
-    const {
-        title,
-        cellKey,
-        mappings,
-        predictionTagsById,
-        onMappingRemoveClick,
-        onMappingAddClick,
-        disabled,
-    } = props;
-
-    const [addInputShown, showAddInput, hideAddInput] = useBooleanState(false);
-    const handleTagRemove = useCallback((tagKey: string) => {
-        onMappingRemoveClick(cellKey, tagKey);
-    }, [
-        cellKey,
-        onMappingRemoveClick,
-    ]);
-
-    const tagRendererParams = useCallback((tagKey: string) => ({
-        className: styles.tag,
-        children: predictionTagsById?.[tagKey]?.name,
-        actions: (
-            <QuickActionButton
-                className={styles.removeButton}
-                name={tagKey}
-                onClick={handleTagRemove}
-                variant="transparent"
-            >
-                <IoCloseOutline />
-            </QuickActionButton>
-        ),
-        variant: 'accent' as const,
-    }), [
-        predictionTagsById,
-        handleTagRemove,
-    ]);
-
-    const filteredMappings = useMemo(() => (
-        mappings?.filter((mapping) => mapping.association.subRowKey === cellKey)
-    ), [mappings, cellKey]);
-
-    const handleTagSelect = useCallback((newTag: string | undefined) => {
-        hideAddInput();
-        if (newTag) {
-            onMappingAddClick(cellKey, newTag);
-        }
-    }, [
-        hideAddInput,
-        onMappingAddClick,
-        cellKey,
-    ]);
-
-    const existingMappings = useMemo(() => (
-        listToMap(
-            filteredMappings,
-            (d) => d.tag,
-            () => true,
-        )
-    ), [filteredMappings]);
-
-    const filteredOptions = useMemo(() => (
-        Object.values(predictionTagsById ?? {}).filter(
-            (tag) => !existingMappings?.[tag.id],
-        )
-    ), [
-        predictionTagsById,
-        existingMappings,
-    ]);
-
-    return (
-        <FrameworkTagRow
-            title={title}
-            rightContent={(
-                <>
-                    <List
-                        data={filteredMappings}
-                        renderer={Tag}
-                        rendererParams={tagRendererParams}
-                        keySelector={mappingKeySelector}
-                    />
-                    {addInputShown && (
-                        <>
-                            <SelectInput
-                                value={undefined}
-                                variant="general"
-                                name={undefined}
-                                options={filteredOptions}
-                                onChange={handleTagSelect}
-                                keySelector={predictionKeySelector}
-                                labelSelector={predictionLabelSelector}
-                                groupKeySelector={predictionGroupKeySelector}
-                                groupLabelSelector={predictionGroupKeySelector}
-                                grouped
-                            />
-                            <QuickActionButton
-                                name={undefined}
-                                onClick={hideAddInput}
-                            >
-                                <IoCloseOutline />
-                            </QuickActionButton>
-                        </>
-                    )}
-                    {!disabled && !addInputShown && (
-                        <QuickActionButton
-                            name={undefined}
-                            onClick={showAddInput}
-                        >
-                            <IoAdd />
-                        </QuickActionButton>
-                    )}
-                </>
-            )}
-        />
-    );
-}
+const matrix1dAssociationKeySelector = (
+    mapping: Matrix1dMappingsItem,
+) => mapping.association.subRowKey;
 
 interface CellItem {
     subRowKey: string;
@@ -181,7 +44,7 @@ interface Props {
     mappings: Matrix1dMappingsItem[] | undefined;
     onMappingsChange: (newMappings: Matrix1dMappingsItem[], widgetPk: string) => void;
     disabled?: boolean;
-    predictionTagsById: Record<string, PredictionTag> | undefined;
+    predictionTags: PredictionTag[] | undefined;
 }
 
 function Matrix1dTagInput(props: Props) {
@@ -191,7 +54,7 @@ function Matrix1dTagInput(props: Props) {
         mappings,
         onMappingsChange,
         disabled,
-        predictionTagsById,
+        predictionTags,
     } = props;
 
     const sortedCells = useMemo(() => (
@@ -276,18 +139,19 @@ function Matrix1dTagInput(props: Props) {
 
     const tagRowRendererParams = useCallback((_: string, cell: CellItem) => ({
         title: cell.subRowLabel,
-        cellKey: cell.subRowKey,
+        itemKey: cell.subRowKey,
         onMappingRemoveClick: handleCellRemove,
         onMappingAddClick: handleCellAdd,
+        associationKeySelector: matrix1dAssociationKeySelector,
         mappings,
-        predictionTagsById,
+        predictionTags,
         disabled,
     }), [
         disabled,
         handleCellRemove,
         handleCellAdd,
         mappings,
-        predictionTagsById,
+        predictionTags,
     ]);
 
     const subRowGroupRendererParams = useCallback((groupKey: string) => ({
@@ -300,7 +164,7 @@ function Matrix1dTagInput(props: Props) {
             className={_cs(className, styles.matrixTagInput)}
             data={sortedCells}
             keySelector={cellKeySelector}
-            renderer={TagRow}
+            renderer={FrameworkTagRow}
             rendererParams={tagRowRendererParams}
             filtered={false}
             pending={false}
